pyproject.toml
[tool.poetry]
name = "kirclub"
version = "0.1.0"
description = ""
authors = ["Saúl Sánchez Rangel <88846065+sauls6@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.2"
whitenoise = "^6.6.0"
pillow = "^10.2.0"
django-ninja = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
